+++ src/providers/ad/ad_machine_pw_renewal.c
@@ -23,7 +23,7 @@
 
 #include "util/util.h"
 #include "util/strtonum.h"
-#include "providers/dp_ptask.h"
+#include "providers/be_ptask.h"
 #include "providers/ad/ad_common.h"
 
 #ifndef RENEWAL_PROG_PATH
