+++ src/providers/dp_refresh.c
@@ -24,7 +24,7 @@
 #include <ldb.h>
 
 #include "providers/dp_backend.h"
-#include "providers/dp_ptask.h"
+#include "providers/be_ptask.h"
 #include "providers/dp_refresh.h"
 #include "util/util_errors.h"
 #include "db/sysdb.h"
