+++ src/providers/ldap/sdap_sudo_refresh.c
@@ -23,7 +23,7 @@
 #include <tevent.h>
 
 #include "util/util.h"
-#include "providers/dp_ptask.h"
+#include "providers/be_ptask.h"
 #include "providers/ldap/sdap_sudo.h"
 #include "providers/ldap/sdap_sudo_shared.h"
 #include "db/sysdb_sudo.h"
